BE/requirements.txt
Django>=3.1
djangorestframework
requests
gunicorn
bs4